Implement Harvested document facet
A very generic "HarvestedRecord" set of metadata has been added to
Nuxeo's default File doctype.  This allows the metadata that's been
retrieved from remote OAI-PMH repositories to be stored on document
records without conflicting with Nuxeo's own metadata elements.

The "Harvested" facet also includes a "rawMetadataContent" field
which is intended to contain the XML file corresponding to the
metadata harvested from the remote host.  This will aid in debugging
harvest mechanisms.

The extension can be tested by installing the
catalog-skateboard-marketplace package onto a Nuxeo host, and
examining a file via the Nuxeo API. For example, if a file `mydoc` is
present at `/Testing/Documents/mydoc` where `Testing` is a workspace
under the "default domain," and `Documents` is a folder inside it,
then an HTTP GET against `http://localhost:8080/nuxeo/site/api/v1/path/default-domain/workspaces/TestingDocuments/mydoc`
with the `X-NXDocumentProperties` header set to "*", then the response
should contain values like

     ...
     "hrv:description": [],
     "hrv:coverage": [],
     "hrv:rawMetadataContent": null,
     ...

indicating that the metadata for the new facet is available on standard
Nuxeo File documents.