Create a standard Nuxeo session from LTI launch
This is a stopgap until #22 can be implemented.  It adds the standard
Nuxeo identity cookie value to the session, using the principal that
the NuxeoOAuthFilter adds to the request before it reaches our
application code.

This will be irritating to do for multiple tools or pathways to
launching an LTI session, so abstracting it into its own
authentication filter is a good long-term plan.

Fixes #103.