Change error handling.
Add code to handle exceptions thrown in finally blocks without losing exceptions
thrown in the corresponding try block.

Previously, when an OAI protocol error was received, an OAIProtocolException
would be thrown, as expected. But, if the OAIResponseHandler threw some other
exception when it was notified that the harvest had ended, this second exception
would be thrown by the harvester and the protocol exception would be lost.

Also make sure MergingOAIResponseHandler doesn't try to finish off the XML if
there's been a harvesting error, because it's not really possible to do so.