Refactor running harvester management into a separate class.
Also stop harvests if we time out when offering records to the queue, and make
clear that each HarvestAgent is single-use. Probably we'll end up making one per
harvest, and we'll definitely not reuse them.

TODO: should we _really_ be shutting the executor down when stopping? I want
to stop all the threads, but I might also want to reuse this pool....