Move scope of guava & freemarker dependencies.
Putting the test scope in the parent pom makes life more difficult than it needs
to be. Example: dropwizard-core has guava as a compile-time dependency, so
executables built using dropwizard stuff naturally expect guava to be available
at runtime. With our poms the way they are, it's necessary to explicitly add
guava as a compile-time dependency to get such executables to run, because
our test dependency overrides dropwizard's compile-time dependency.

Same story for freemarker.

The achievement of the civilization.