Use javascript for the second DynamoDB lambda function.
The java version failed with mysterious jackson deserializing errors. Not worth
trying to figure out what's wrong -- we'll just use javascript instead.

This module use the frontend maven plugin to download node and npm, install
dependencies, and lint the file. A shell script is provided to upload the
zip file resulting from the build to S3.#