Rename `verifyNewItem` to `checkNewItem`.
The other verify functions and this one are fundamentally different:
`verifyEvent` and `verifyQueues` both perform a sequence of checks
where, if one fails, subsequent checks will also fail (or worse, throw),
whereas `verifyNewItem` performs a sequence of checks, all of which
should be performed. 

Renaming hopefully makes it clear that these are two distinct types of
checks and eliminates a source of confusion.