Lots of changes.
* Refactored `Tests` utility class. New `WireMock` class contains all the
wiremock things.

* Refactored MergingOAIResponseHandler.
It now takes a delegate `OAIEventHandler` rather than an `OutputStream`, making it
more flexible (one could, e.g., buffer and/or filter `XMLEvents` in this
delegate).

* Add `OAIEventHandler.close()` method.
It turns out to be necessary if we want more abstraction. Interface's default
implementation is empty.

* Don't swallow `XMLStreamExceptions` thrown when `XMLEventWriters` are closed.
I'm not sure why I thought that was a good idea. Possibly because it simplifies
exception handling when closing things down. A better solution is to use the
`suppressExceptions` method of `Harvester` when closing. Thus `Functions` was
born.