Add StAX JUnit rule for testing.
StAXRule will run tests repeatedly, once for each specified StAX
implementation. All three that I know/care about (JDK, Xerces, and Woodstox)
are supported.

One simply adds the rule to a test class, and all test methods in that class
will be run with the implementations specified in the constructor. Unlike some
other rules (e.g., ExpectedException), this rule cannot react to changes made
to the rule object within the test methods, because the implementations to
use must be determined and set up before the methods are called. As a workaround
for tests that must use a different set of implementations to those specified in
the rule's constructor, the @StAXImplementations method annotation is provided.
The set of implementations specified in this annotation take precedence over
those specified in the rule itself.

This eliminates the need to remember to run the tests with the various
maven profiles, and ensures that CI builds will find problems with
_any_ of the various StAX implementations. The maven profiles remain,
however, (with slight modification) so that we can use either of the
non-JDK implementations in the actual service, should we choose.

The implementation could use a bit of touching up still. Not sure this is
even its final form.