Refactored MergingOAIResponseHandler. Handle XMLEventWriter.close() exceptions.
* Merging handler now takes a delegate `OAIEventHandler` rather than
an `OutputStream`, making it more flexible (one could, e.g., buffer
and/or filter `XMLEvents` in this delegate). This requires the
addition of a `close()` method to `OAIEventHandler`.

* Don't swallow `XMLStreamExceptions` thrown when `XMLEventWriters`
are closed.  I'm not sure why I thought that was a good idea. Possibly
because it simplifies exception handling when closing things down. A
better solution is to use the `suppressExceptions` method of
`Harvester` when closing. Thus `Functions` was born.