Rework migrations considerably.
Previously the migrations contained SQL-formatted changesets, some of
which included `CREATE ALIAS` commands. Aliases typically refer to
Java classes (by their full name). Thus, refactoring of the Java code
actually implementing database functions must include a migration
altering the alias.

Unfortunately, if the original `CREATE ALIAS` commands remain in the set of
migrations to perform, the database can no longer be created from scratch
with the refactored code in place (because the classes/functions originally
referred to are missing/have moved/have a different scope). Simply removing
them causes problems when running migrations against existing databases,
and attempts to skip them conditionally have failed in one way or another.

So, here's the plan: start over. The existing database will be wiped
on the next deploy, and a fresh copy created. Aliases are now handled
separately, by a separate SQL file that runs first on every migration
(but only if the file has changed since the previous migration). This
script drops all existing aliases and creates only those that are
meant to exist in the current database revision.

Since SQL-formatted changesets appear to have limited support (in
particular, most of the preconditions are not supported), we'll avoid
them in the future. Where SQL is required or convenient, we'll include
it via `sql` or `sqlFile` in a YAML changeset.

Lesson learned.