Fix config and migrations.
Apparently dropwizard-java8 neglected to actually support the old format with
the new class, so config must be `PT<duration>` instead of just `<duration>`.

The secret to making migrations work both from the tests/IDE and from
the command line via the jar is to use `sql/foo.sql` and make the
thing not relative to the changelog file, apparently.