Add checking for invalid repository base URIs.
The CREATE_JOB stored function now verifies that repositories
exist. If any harvests in the job specify an invalid URI (one which we
don't know about, i.e., it's not in our list of repositories), the
transaction is rolled back. The resulting JobInfo object now includes
information about harvests, including invalid ones for error reporting
purposes. This eliminates an HTTP 500 when sending requests for
invalid base URIs, which was annoying.

The real purpose behind the changes, though, is to get more info about
the harvest into the tags so that we can report on failed harvests
easily. New stuff in tags: repository institution and name.