Avoid race condition with pipe connection.
Because the input and output pipes were connected in the child thread
(to avoid handling an exception in the parent), there was a race
condition. It was possible for the database update to try to read from
the input end (in the parent thread) before the connection was made in
the child thread.