Add cooldown periods for provisioned write capacity increase and decrease.
The monitor will not try to increase capacity if capacity has been
increased within the increase cooldown period. It will not attempt to
decrease capacity if capacity has been adjusted either upward or
downward during the last decrease cooldown period.

The cooldown for decrease is necessarily much larger than that for
increase, because the number of decrease requests is limited to 4 per
day.